# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: SwiftEats Food Delivery Platform API
  description: |
    Comprehensive API specification for the SwiftEats real-time food delivery platform.
    
    ## Overview
    SwiftEats is a scalable, resilient, and high-performance microservices-based food delivery platform.
    
    ## Key Features
    - Real-time GPS tracking for drivers
    - High-performance order processing (500 orders/minute)
    - Real-time analytics (2,000 GPS events/second)
    - Multi-layer caching for sub-200ms response times
    
    ## Services
    - Auth Service (Port: 3001) - Authentication and authorization
    - User Service (Port: 3002) - User management
    - Restaurant Service (Port: 3003) - Restaurant and menu management
    - Order Service (Port: 3004) - Order processing
    - Delivery Service (Port: 3005) - Delivery tracking
    - Notification Service (Port: 3006) - Notifications
    - Payment Service (Port: 3007) - Payment processing
    - Analytics Service (Port: 3008) - Real-time analytics
    
    ## Authentication
    All API endpoints require JWT authentication unless marked as public.
    Include the JWT token in the Authorization header: `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Standard endpoints: 1000 requests per minute
    - Order endpoints: 100 requests per minute
    - Analytics endpoints: 5000 requests per minute
    
    ## Health Checks
    Each service provides a minimal health endpoint:
...
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: "Bad Request"
        error:
          type: string
          example: "Validation failed"
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: "/api/v1/orders"
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer

    # Auth Service schemas
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 6
          example: "password123"
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        firstName:
          type: string
          minLength: 2
        lastName:
          type: string
          minLength: 2
        phone:
          type: string
        address:
          type: string

    # User Service schemas
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        address:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        restaurantId:
          type: string
          format: uuid
        totalAmount:
          type: number
          format: float
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, picked_up, delivered, cancelled]
        deliveryAddress:
          type: string
        deliveryInstructions:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        menuItemId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: float
        totalPrice:
          type: number
          format: float
        specialInstructions:
          type: string
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Refresh the access token using a refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Invalid refresh token

  # User Service Endpoints
...
  /orders:
    get:
      tags:
        - Orders
      summary: List user orders
      description: Retrieve orders for the current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, picked_up, delivered, cancelled]
          description: Filter by order status
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order details
      description: Retrieve detailed information about a specific order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

  # Delivery Service Endpoints
...